remove FedoraObjectHandler interface and VersionAbstractionFedoraObjectHandler.java
