+++ src/main/java/org/fcrepo/migration/handlers/ObjectAbstractionStreamingFedoraObjectHandler.java
@@ -19,12 +19,14 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Collections;
 
 import org.fcrepo.migration.DatastreamVersion;
-import org.fcrepo.migration.FedoraObjectHandler;
 import org.fcrepo.migration.ObjectInfo;
 import org.fcrepo.migration.ObjectProperties;
 import org.fcrepo.migration.ObjectReference;
+import org.fcrepo.migration.ObjectVersionReference;
+import org.fcrepo.migration.FedoraObjectVersionHandler;
 import org.fcrepo.migration.StreamingFedoraObjectHandler;
 
 /**
@@ -36,7 +38,7 @@ import org.fcrepo.migration.StreamingFedoraObjectHandler;
  */
 public class ObjectAbstractionStreamingFedoraObjectHandler implements StreamingFedoraObjectHandler {
 
-    private FedoraObjectHandler handler;
+    private FedoraObjectVersionHandler versionHandler;
 
     private ObjectInfo objectInfo;
 
@@ -50,10 +52,10 @@ public class ObjectAbstractionStreamingFedoraObjectHandler implements StreamingF
 
     /**
      * the object abstraction streaming fedora object handler.
-     * @param objectHandler the fedora object handler
+     * @param versionHandler the fedora object version handler
      */
-    public ObjectAbstractionStreamingFedoraObjectHandler(final FedoraObjectHandler objectHandler) {
-        this.handler = objectHandler;
+    public ObjectAbstractionStreamingFedoraObjectHandler(final FedoraObjectVersionHandler versionHandler) {
+        this.versionHandler = versionHandler;
         this.dsIds = new ArrayList<String>();
         this.dsIdToVersionListMap = new HashMap<String, List<DatastreamVersion>>();
     }
@@ -86,33 +88,92 @@ public class ObjectAbstractionStreamingFedoraObjectHandler implements StreamingF
 
     @Override
     public void completeObject(final ObjectInfo object) {
+        final var objectReference = new ObjectReference() {
+            @Override
+            public ObjectInfo getObjectInfo() {
+                return objectInfo;
+            }
+
+            @Override
+            public ObjectProperties getObjectProperties() {
+                return objectProperties;
+            }
+
+            @Override
+            public List<String> listDatastreamIds() {
+                return dsIds;
+            }
+
+            @Override
+            public List<DatastreamVersion> getDatastreamVersions(final String datastreamId) {
+                return dsIdToVersionListMap.get(datastreamId);
+            }
+
+            @Override
+            public boolean hadFedora2Disseminators() {
+                return disseminatorsSkipped > 0;
+            }
+        };
         try {
-            handler.processObject(new ObjectReference() {
-                @Override
-                public ObjectInfo getObjectInfo() {
-                    return objectInfo;
-                }
+            final Map<String, List<DatastreamVersion>> versionMap = buildVersionMap(objectReference);
+            final List<String> versionDates = new ArrayList<String>(versionMap.keySet());
+            Collections.sort(versionDates);
+            final List<ObjectVersionReference> versions = new ArrayList<ObjectVersionReference>();
+            for (final String versionDate : versionDates) {
+                versions.add(new ObjectVersionReference() {
+                    @Override
+                    public ObjectReference getObject() {
+                        return objectReference;
+                    }
 
-                @Override
-                public ObjectProperties getObjectProperties() {
-                    return objectProperties;
-                }
+                    @Override
+                    public ObjectInfo getObjectInfo() {
+                        return objectReference.getObjectInfo();
+                    }
 
-                @Override
-                public List<String> listDatastreamIds() {
-                    return dsIds;
-                }
+                    @Override
+                    public ObjectProperties getObjectProperties() {
+                        return objectReference.getObjectProperties();
+                    }
 
-                @Override
-                public List<DatastreamVersion> getDatastreamVersions(final String datastreamId) {
-                    return dsIdToVersionListMap.get(datastreamId);
-                }
+                    @Override
+                    public String getVersionDate() {
+                        return versionDate;
+                    }
 
-                @Override
-                public boolean hadFedora2Disseminators() {
-                    return disseminatorsSkipped > 0;
-                }
-            });
+                    @Override
+                    public List<DatastreamVersion> listChangedDatastreams() {
+                        return versionMap.get(versionDate);
+                    }
+
+                    @Override
+                    public boolean isLastVersion() {
+                        return versionDates.get(versionDates.size() - 1).equals(versionDate);
+                    }
+
+                    @Override
+                    public boolean isFirstVersion() {
+                        return versionDates.get(0).equals(versionDate);
+                    }
+
+                    @Override
+                    public int getVersionIndex() {
+                        return versionDates.indexOf(versionDate);
+                    }
+
+                    @Override
+                    public boolean wasDatastreamChanged(final String dsId) {
+                        for (final DatastreamVersion v : listChangedDatastreams()) {
+                            if (v.getDatastreamInfo().getDatastreamId().equals(dsId)) {
+                                return true;
+                            }
+                        }
+                        return false;
+                    }
+
+                });
+            }
+            versionHandler.processObjectVersions(versions);
         } finally {
             cleanForReuse();
         }
@@ -130,5 +191,21 @@ public class ObjectAbstractionStreamingFedoraObjectHandler implements StreamingF
     private void cleanForReuse() {
         this.dsIds.clear();
         this.dsIdToVersionListMap.clear();
+    }
+
+    private Map<String, List<DatastreamVersion>> buildVersionMap(final ObjectReference object) {
+        final Map<String, List<DatastreamVersion>> versionMap = new HashMap<String, List<DatastreamVersion>>();
+        for (final String dsId : object.listDatastreamIds()) {
+            for (final DatastreamVersion v : object.getDatastreamVersions(dsId)) {
+                final String date = v.getCreated();
+                List<DatastreamVersion> versionsForDate = versionMap.get(date);
+                if (versionsForDate == null) {
+                    versionsForDate = new ArrayList<DatastreamVersion>();
+                    versionMap.put(date, versionsForDate);
+                }
+                versionsForDate.add(v);
+            }
+        }
+        return versionMap;
     }
 }
