+++ src/test/resources/spring/ocfl-user-it-setup.xml
@@ -64,10 +64,6 @@
     </bean>
 
     <bean id="objectAbstraction" class="org.fcrepo.migration.handlers.ObjectAbstractionStreamingFedoraObjectHandler">
-        <constructor-arg ref="versionAbstraction"/>
-    </bean>
-
-    <bean id="versionAbstraction" class="org.fcrepo.migration.handlers.VersionAbstractionFedoraObjectHandler">
         <constructor-arg ref="minimal"/>
     </bean>
