+++ src/test/resources/spring/version-abstraction.xml
@@ -23,10 +23,6 @@
     </bean>
 
     <bean id="objectAbstraction" class="org.fcrepo.migration.handlers.ObjectAbstractionStreamingFedoraObjectHandler">
-        <constructor-arg ref="versionAbstraction"/>
-    </bean>
-
-    <bean id="versionAbstraction" class="org.fcrepo.migration.handlers.VersionAbstractionFedoraObjectHandler">
         <constructor-arg ref="versionHandler"/>
     </bean>
